refactor(main): register IPC per domain, and move guidance next to what it guards
`src/main/index.ts` was 2,697 lines holding 120 `handleTrusted` registrations
inline, interleaved with window and lifecycle code, so reading one domain's
trust-boundary surface meant scrolling past every other. It is now 1,627 lines,
and `registerIpcHandlers` is a dispatcher: sixteen `register*IpcHandlers` calls
against fifteen new files under `src/main/ipc/`, following the shape
`register-team-handlers.ts` already established for one domain and stopped at.
The registrations, their input parsers and their route helpers moved unchanged.

The largest of them is `register-agent-handlers.ts` at 431 lines - bots, the
conversation and the queue, which are one surface. Memories, routines and
attachments came out of it into their own registrars: each is a closed
sub-surface with its own parsers and no shared state with the rest, and leaving
all four in one file would have reproduced index.ts at a smaller scale.

`registerIpcHandlers` takes a named `IpcHandlerDependencies` object rather than
21 positional parameters. Three of those were `setupFile`,
`analyticsPreferenceFile` and `updatePreferenceFile` - three paths of the same
type, where swapping two compiles cleanly and writes one preference into
another's file, which nothing would catch.

Root AGENTS.md went 279 lines to 202. Renderer UI, Database migrations and Team
API protocol compatibility now live in `src/renderer/AGENTS.md`,
`src/backend/AGENTS.md` and `packages/contracts/AGENTS.md`, which load when you
open a file under them - where a rule about migrations is worth more than in a
file read once at the top of a session. The D1 deploy-race rule went to a new
`apps/auth-api/AGENTS.md` rather than to `src/backend`, which is not an ancestor
of `apps/auth-api/migrations/` and would not have loaded for the change it
governs. `src/main/AGENTS.md` is new and is the only place that
rule is written down: an endpoint is declared in `packages/contracts` and
registered in `src/main/ipc/`, never inline in index.ts. Each of the fifteen
modules opens with what its domain is and nothing about policy - the directory's
AGENTS.md loads when you open one of them, so a copy in every header is fifteen
things to keep in sync with it.

Each of the five also gains a waiting table, which is what Tests rule 4 has been
missing. The rule says a test that needs a timeout is wrong; nothing said what
to use instead, so a rejected `setTimeout` becomes a longer `vi.waitFor` timeout
rather than a real barrier. The tables name the barriers each directory already
has - `waitFor(manager, predicate)`, `nextRoutinesChanged`, the `emit*` bridges
in `app-test-harness.ts`, `findByRole`, a `server.listen` callback.

One test was added, to `ipc-channel-coverage.test.ts`: every registrar under
`src/main/ipc` must be called exactly once by the dispatcher. Splitting index.ts
is what makes that reachable-but-unwired state possible, and nothing else sees
it - the coverage scan reads sources, so an orphaned registrar's channels still
count as registered while every one of them rejects at runtime. `tsc` covers
only half: deleting the call alone is TS6133 since the call site is the import's
only use, but deleting the import with it exits 0, and so does a new registrar
nobody wired up. Verified by removing both the call and the import for
`registerMemoryIpcHandlers`, which leaves `typecheck:node` green and fails the
new assertion naming the registrar and its file.

Nothing else was added. Three properties that look like they need a test are
already mechanical, and each was verified by breaking it:

- `ipc-channel-coverage.test.ts` fails "registers each request channel exactly
  once" if a moved block is left behind in both files. Duplicating one
  `agentListMemories` registration reported both paths by name; that assertion
  is what makes this move safe.
- it already fails on a bare `ipcMain` anywhere in `src/main` outside the
  wrapper module, which is what keeps the fifteen new files inside the sender
  check.
- a method missing from `mock-openbot.ts` is TS2741, and one the interface never
  declared is TS2353, since both it and the preload bridge are annotated
  `: OpenBotDesktopApi`. This is the "fourth mirror" half of #174 item 6, and
  the issue itself files generating that mirror as a separate decision.
